@@ -1,13 +1,12 @@
 name: mongo_dart_query
-version: 0.4.2
+version: 1.0.0-nullsafety
 description: Query builder for mongo_dart and objectory
 homepage: https://github.com/mongo-dart/mongo_dart_query
 environment:
-  sdk: ">=2.5.2 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 dependencies:
-  bson: "^0.3.2"
-  meta: ^1.1.7
+  bson: ^1.0.0-nullsafety
+  meta: ^1.3.0-nullsafety.6
 dev_dependencies:
-  test: ^1.14.0
-  pedantic: ^1.9.0
-  #dart_coveralls: ">=0.4.0"
+  test: ^1.16.0-nullsafety.9
+  pedantic: ^1.10.0-nullsafety.3
